requirements.txt
# Core ML
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
hmmlearn>=0.3.0
scipy>=1.11.0
joblib>=1.3.0

# Data
yfinance>=0.2.28
pyarrow>=13.0.0

# Options IV (optional — fallback to Yahoo IV field)
# py_vollib>=1.0.1

# Portfolio Optimization
pypfopt>=1.5.5

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# LLM
anthropic>=0.39.0

# Plotting (for training diagnostics)
matplotlib>=3.7.0

# Utilities
python-dotenv>=1.0.0
